requirements.txt
# === Core scientific stack ===
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0

# === Graph & mapping ===
osmnx>=1.9.0
networkx>=3.2.0
shapely>=2.0.0
pyproj>=3.6.0

# === Deep learning ===
torch>=2.3.0
torch-geometric>=2.5.0

# === API & backend ===
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
requests>=2.32.0
aiohttp>=3.10.0

# === Visualization & UI ===
folium>=0.16.0
streamlit>=1.36.0

# === Traffic & Live data ===
geopandas>=1.0.0
ratelimit>=2.2.1
tenacity>=8.3.0  # retry logic for API calls
cachetools>=5.3.3  # caching TomTom API responses

# === Optional for VRP & evaluation ===
ortools>=9.10.0
matplotlib>=3.9.0
tensorboard>=2.17.0

# === Dev utilities ===
pytest>=8.2.0
black>=24.4.0
isort>=5.13.0
